move zeros to its own file
