+++ src/Flux.jl
@@ -27,6 +27,7 @@ using CuArrays
 const use_cuda = Ref(false)
 
 include("utils.jl")
+include("zeros.jl")
 include("onehot.jl")
 include("functor.jl")
